requirements.txt
alpaca-trade-api==0.46
alpha-vantage==2.1.3
altgraph==0.17
appdirs==1.4.3
arcade==2.3.10
arrow==0.15.6
asyncio-nats-client==0.10.0
attrs==19.3.0
bcrypt==3.1.7
beautifulsoup4==4.9.0
black==19.10b0
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
chromedriver-binary==83.0.4103.14.0
click==7.1.1
dataclasses==0.7
Deprecated==1.2.7
fastapi==0.53.2
furl==2.1.0
gunicorn==20.0.4
h11==0.9.0
httplib2==0.17.3
httptools==0.1.1
idna==2.9
iexfinance==0.4.3
joblib==0.14.1
lxml==4.5.0
multitasking==0.0.9
myModule==1.0.0
numpy==1.18.2
orderedmultidict==1.0.1
pandas==1.0.3
passlib==1.7.2
pathspec==0.8.0
Pillow==7.1.0
pycparser==2.20
pydantic==1.4
pyglet==1.5.2
PyInstaller==3.6
PyJWT==1.7.1
pyserial==3.4
python-dateutil==2.8.1
python-dotenv==0.13.0
python-env==1.0.0
python-multipart==0.0.5
pytiled-parser==0.9.3
pytz==2019.3
regex==2020.4.4
requests==2.23.0
schedule==0.6.0
scikit-learn==0.22.2.post1
scipy==1.4.1
selenium==4.0.0a5
six==1.14.0
sklearn==0.0
sms==0.4
smsc==0.1.1
soupsieve==2.0
SQLAlchemy==1.3.16
starlette==0.13.2
textmagic==2.0.3
tiingo==0.12.0
toml==0.10.0
typed-ast==1.4.1
typing==3.7.4.1
urllib3==1.24.3
uvicorn==0.11.3
uvloop==0.14.0
websocket-client==0.57.0
websockets==8.1
wrapt==1.12.1
xgboost==1.0.2
yfinance==0.1.54
